renamed test_util files to match namespaces exactly
