+++ src/thunderbots/software/CMakeLists.txt
@@ -235,7 +235,7 @@ if (CATKIN_ENABLE_TESTING)
 
     catkin_add_gtest(world_test
             test/world/world.cpp
-            test/test_util/util.cpp
+            test/test_util/test_util.cpp
             ai/world/world.cpp
             ai/world/ball.cpp
             ai/world/field.cpp
@@ -245,8 +245,8 @@ if (CATKIN_ENABLE_TESTING)
     target_link_libraries(world_test ${catkin_LIBRARIES})
 
     catkin_add_gtest(test_util_test
+            test/test_util/test_util_test.cpp
             test/test_util/test_util.cpp
-            test/test_util/util.cpp
             ai/world/world.cpp
             ai/world/ball.cpp
             ai/world/field.cpp
